fix: Clean up zookeeper (#5)
The Zookeeper dependency is required for a single test method only. This PR moves all ZK* classes to src/test, and marks the zookeeper dependency as test-coped.

Unfortunately, recent versions of zookeeper drag in a logback dependency, that appears to turn on global DEBUG logging. So this PR also exludes those logback dependencies.

Fixes kafka-ops#548